chapter8/pyproject.toml
[project]
name = "chapter8"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain>=1.2.6",
    "langchain-google-genai>=4.2.0",
    "python-dotenv>=1.2.1",
]
